TG-backend/requirements.txt
pymongo==3.12.3
motor==2.5.1
fastapi==0.75.0
fastapi-utils==0.2.1
PyJWT==2.3.0
pydantic==1.9.0
Werkzeug
python-dotenv==0.19.2
pydantic==1.9.0
toml==0.10.2
requests==2.24.0
typing-extensions==3.7.4.3
uvicorn==0.11.8
aiohttp
aiohttp[speedups]